refactor: typerom entity 정의
