+++ src/entity/Location.ts
@@ -1,0 +1,23 @@
+import {
+    BaseEntity,
+    Entity, 
+    PrimaryGeneratedColumn, 
+    Column,
+    CreateDateColumn,
+    UpdateDateColumn,
+} from "typeorm";
+
+@Entity()
+export class Location extends BaseEntity {
+    @PrimaryGeneratedColumn()
+    id!: number;
+
+    @Column({ nullable: true })
+    name!: string;
+
+    @CreateDateColumn({ name: 'created_at' })
+    createdAt!: Date;
+
+    @UpdateDateColumn({ name: 'updated_at' })
+    updatedAt!: Date;
+}
