+++ src/entity/MatchUserApplication.ts
@@ -1,0 +1,27 @@
+import {
+    BaseEntity,
+    Entity, 
+    PrimaryGeneratedColumn, 
+    Column,
+    CreateDateColumn,
+    UpdateDateColumn,
+} from "typeorm";
+import { ApplicationStatus } from "util/Enums";
+
+@Entity()
+export class MatchUserApplication extends BaseEntity {
+    @PrimaryGeneratedColumn()
+    id!: number;
+
+    @Column()
+    quota!: number;
+    
+    @Column({ enum: ApplicationStatus })
+    status!: ApplicationStatus;
+
+    @CreateDateColumn({ name: 'created_at' })
+    createdAt!: Date;
+
+    @UpdateDateColumn({ name: 'updated_at' })
+    updatedAt!: Date;
+}
