+++ src/util/Enums.ts
@@ -1,0 +1,25 @@
+export enum Level {
+    LOW = "LOW",
+    MIDDLE = "MEDIUM",
+    HIGH = "HIGH",
+}
+
+export enum MatchType {
+    SOCCER = "SOCCER",
+    FUTSAL5 = "FUTSAL5", 
+    FUTSAL6 = "FUTSAL6",
+}
+
+export enum MatchStatus {
+    WAITING = "WAITING",
+    CONFIRMED = "CONFIRMED",
+    DENIED = "DENIED",
+    CANCEL = "CANCEL",
+}
+
+export enum ApplicationStatus {
+    WAITING = "WAITING",
+    CONFIRMED = "CONFIRMED",
+    DENIED = "DENIED",
+    CANCEL = "CANCEL",
+}
